prod email

+++ db/employees.js
@@ -1,10 +1,16 @@
-
 const mysqlConnection  = require('../database.js');
+
+// console.log('basedatos')
+const key_rnd = Math.floor(Math.random() * 16777215).toString(16)
+let acceso = true
+/* let num = '8b38afwid'
+let acceso = true */
+
 
 const get_employees = async (req, res) => {
   mysqlConnection.query('SELECT * FROM employees', (err, rows, fields) => {
     if(!err) {
-      console.log('res ',rows)
+      // console.log('res ',rows)
       res.json(rows);
     } else {
       console.log(err);
@@ -26,49 +32,91 @@ const get_employee = async (req, res) => {
 }
 
 const create_employee = async (req, res) => {
+  const { key } = req.params;
+
+  // console.log('create ')
+  // console.log('create rnd y key  ', key_rnd, key)
+  // console.log('create name = ', req.body.name)
+
+
   const sql = 'insert into employees set ?'
   const customerObj = {
     name: req.body.name,
     salary: req.body.salary
   }
-  mysqlConnection.query(sql, customerObj, (err, rows) => {
-    if(!err) {
-      res.send('Employeed Saved');
+
+  // console.log('acceso ', acceso)
+
+  if (acceso) {
+
+    // console.log('access ok ')
+    acceso = false
+    if (key_rnd == req.params.key) {
+
+      // console.log('code ok ', req.params.key)
+      
+
+      mysqlConnection.query(sql, customerObj, (err, rows) => {
+        if(!err) {
+          res.send('Employeed Saved');
+        } else {
+          console.log(err);
+        }
+      });
+
     } else {
-      console.log(err);
+      // console.log('not match code')
+      res.send('not match code');
     }
-  });
+
+  } else {
+    // console.log('not access')
+    res.send('not access');
+    
+  }
+
+  
 }
 
 const update_employee = async (req, res) => {
+
+  
   const { name, salary } = req.body; // argumentos
   const { id } = req.params; // parametros
+
   /* const customerObj = {
     name: req.body.name,
     salary: req.body.salary
   } */
-  console.log('nmae ', name, 'salary ', salary) // ok
+  // console.log('name ', name, 'salary ', salary) // ok
   const sql = `update employees set name='${name}', salary='${salary}' where id=${id}`
-  mysqlConnection.query(sql, [ name, salary ], (err, rows) => {
-    if(!err) {
-      res.send('Employeed Saved');
-    } else {
-      throw err
-      console.log(err);
-    }
-  });
-  res.send('Update employee')
+
+  if (id > 2) {
+    mysqlConnection.query(sql, [ name, salary ], (err, rows) => {
+      if(!err) {
+        // res.send('Employeed Updated');
+        res.json({status: 'Employee Updated'});
+      } else {
+        throw err
+        console.log(err);
+      }
+    });
+    // res.send('Update employee')
+  }
 }
 
 const delete_employee = async (req, res) => {
   const { id } = req.params;
-  mysqlConnection.query('DELETE FROM employees WHERE id = ?', [id], (err, rows, fields) => {
-    if(!err) {
-      res.json({status: 'Employee Deleted'});
-    } else {
-      console.log(err);
-    }
-  });
+
+  if (id > 5) {
+    mysqlConnection.query('DELETE FROM employees WHERE id = ?', [id], (err, rows, fields) => {
+      if(!err) {
+        res.json({status: 'Employee Deleted'});
+      } else {
+        console.log(err);
+      }
+    });
+  }
 }
 
-module.exports = { get_employees, get_employee, create_employee, update_employee, delete_employee }
+module.exports = { get_employees, get_employee, create_employee, update_employee, delete_employee, key_rnd }
